Add commented code for refactor to Context
Refactored code being used in Main.js and NoteList.js to use context.
This required movement of the getNotesById function into the NoteList as
opposed to the Main component.

Created a generateNotesList function that accepts a list of notes to
generate the NoteListItems need for the component to display the notes
necessary.

This refactor is to ensure the app still displays all notes when a
folder is not selected by if a folder is selected only notes tied to
that folder should be displayed.